requirements.txt
numpy==1.22.4
loguru
tqdm
pyyaml
pandas
openpyxl
matplotlib
scikit-learn
scikit-image
scipy
opencv-python
seaborn
albumentations==1.1.0
tabulate
warmup-scheduler
transformers==4.21.3
torchprofile
torchmetrics
einops
ptflops
torchsummary
torchsummaryx
segmentation-mask-overlay==0.3.4
timm==0.6.12
tifffile
pillow
thop
# Windows-friendly pinned wheel; unpinned latest may try to build from source.
simpleitk==2.3.1
nibabel
h5py
huggingface-hub==0.11.0
ml_collections
tensorboardx
medpy
